Add README and move it so that the import paths are generic
